requirements/base.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.2.1
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==7.6.10
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
FastAPI-SQLAlchemy==0.2.1
flake8==7.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isort==6.0.0
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyflakes==3.2.0
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.0.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
stripe==11.5.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
